context вынесен в отдельный компонент AppContext ContentOverlay вынесен в отдельный компонент
