+++ src/components/ChatList/ChatList.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import Context from '../../context/context';
+import { Context } from '../../context/AppContext';
 
 import Chat from '../Chat/Chat';
 import './ChatList.css';
